requirements.txt
flask
pywebview
torch
transformers
numpy
pandas
scikit-learn
textblob
joblib